# Documentazione/Progettazione/openapi.yaml
openapi: 3.0.3
info:
  title: Monopoly IoT API
  version: 1.0.0
  description: |
    API REST esposte dall'API Gateway del sistema Monopoly IoT.

    Il sistema è composto da 5 microservizi:
    - **api-gateway** (porta 8080) — punto di ingresso unico
    - **user-service** (porta 8081) — gestione utenti
    - **match-service** (porta 8082) — partite, tornei, motore di gioco
    - **board-service** (porta 8083) — sedi, tavoli, zone RFID
    - **iot-gateway** (porta 8085) — acquisizione eventi MQTT, buffer offline

    Tutte le richieste vengono inoltrate dal Gateway ai servizi interni.
servers:
  - url: http://localhost:8080
    description: API Gateway locale
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Elenco utenti
      operationId: listUsers
      responses:
        "200":
          description: Lista utenti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags: [Users]
      summary: Crea utente (registrazione)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
            example:
              username: "mario_rossi"
              email: "mario@example.com"
              passwordHash: "abc123"
              fullName: "Mario Rossi"
              active: true
              wantsTournament: false
      responses:
        "200":
          description: Utente creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/users/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Users]
      summary: Dettaglio utente
      operationId: getUserById
      responses:
        "200":
          description: Utente trovato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Utente non trovato
    put:
      tags: [Users]
      summary: Aggiorna utente
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Utente aggiornato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Utente non trovato
    delete:
      tags: [Users]
      summary: Elimina utente
      operationId: deleteUser
      responses:
        "200":
          description: Utente eliminato
        "404":
          description: Utente non trovato
  /api/users/{id}/profile:
    put:
      tags: [Users]
      summary: Aggiorna profilo utente (bio, avatar, genere, età)
      operationId: updateUserProfile
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Profilo aggiornato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Utente non trovato
  /api/users/{id}/wants-tournament:
    put:
      tags: [Users]
      summary: Imposta disponibilità ai tornei
      operationId: setWantsTournament
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: wants
          in: query
          required: true
          schema:
            type: boolean
          description: true = disponibile a partecipare ai tornei
      responses:
        "200":
          description: Preferenza aggiornata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/users/{id}/notice:
    put:
      tags: [Users]
      summary: Imposta comunicazione amministrativa per utente
      operationId: setAdminNotice
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notice]
              properties:
                notice:
                  type: string
                  example: "Il tuo account è stato promosso a Venue Manager."
      responses:
        "200":
          description: Notice impostata
    delete:
      tags: [Users]
      summary: Cancella comunicazione amministrativa
      operationId: deleteAdminNotice
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Notice cancellata
  /api/users/{id}/change-password:
    put:
      tags: [Users]
      summary: Cambia password utente
      operationId: changePassword
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Password modificata con successo
        "400":
          description: Password corrente errata o richiesta non valida

  # ─────────────────────────────────────────────────────────────
  # PLAYER DASHBOARD (→ match-service :8082, routing speciale)
  # ─────────────────────────────────────────────────────────────
...
  /api/venues:
    get:
      tags: [Board]
      summary: Elenco sedi
      operationId: listVenues
      responses:
        "200":
          description: Lista sedi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      tags: [Board]
      summary: Crea sede
      operationId: createVenue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Venue"
      responses:
        "200":
          description: Sede creata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
...
  /api/venues/dashboard/{userId}:
    get:
      tags: [Board]
      summary: Dashboard sede per gestore
      description: |
        Restituisce stato dei tavoli, partite attive, occupancy e ultimi eventi
        per la sede associata all'utente (ownerUserId).
        **Nota**: questa rotta è gestita dal match-service.
      operationId: getVenueDashboard
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Dashboard sede
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDashboardDTO"
        "404":
          description: Sede non trovata per questo gestore
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Identificatore univoco della risorsa
  schemas:
    User:
      type: object
      description: Utente della piattaforma (giocatore, venue manager o admin)
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        username:
          type: string
          example: "mario_rossi"
        email:
          type: string
          format: email
          example: "mario@example.com"
        passwordHash:
          type: string
          description: Hash della password (non restituire in produzione)
        status:
          type: string
          example: "ONLINE"
        fullName:
          type: string
          example: "Mario Rossi"
        active:
          type: boolean
          default: true
        wantsTournament:
          type: boolean
          default: false
          description: true se il giocatore è disponibile per tornei
        bio:
          type: string
        gender:
          type: string
        age:
          type: integer
        avatar:
          type: string
          description: URL o base64 dell'immagine avatar
        adminNotice:
          type: string
          description: Comunicazione mostrata al prossimo accesso
    Venue:
      type: object
      description: Sede fisica dove si trovano i tavoli Monopoly IoT
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          example: "Sala Giochi Centrale"
        location:
          type: string
        address:
          type: string
        city:
          type: string
          example: "Milano"
        type:
          type: string
          description: Tipo di sede (es. "Sala Giochi", "Bar", "Circolo")
        contactInfo:
          type: string
        ownerUserId:
          type: integer
          format: int64
          description: userId del Venue Manager responsabile della sede
...
    VenueDashboardDTO:
      type: object
      description: Dashboard dello stato operativo di una sede
      properties:
        venueName:
          type: string
        matchesToday:
          type: integer
        matchesTrend:
          type: integer
          description: Variazione percentuale rispetto al giorno precedente
        topBoardName:
          type: string
          description: Nome del tavolo con più occupancy
        topBoardOccupancy:
          type: integer
          description: Occupancy 0-100 del tavolo più attivo
        activeBoards:
          type: array
          items:
            $ref: "#/components/schemas/BoardStatusDTO"
        liveEvents:
          type: array
          items:
            $ref: "#/components/schemas/LiveEventDTO"
    BoardStatusDTO:
      type: object
      properties:
        boardName:
          type: string
        status:
          type: string
          enum: [IN_MATCH, WAITING, ONLINE, OFFLINE]
        currentPlayers:
          type: integer
        maxPlayers:
          type: integer
        duration:
          type: string
        matchId:
          type: integer
          format: int64
          nullable: true
    LiveEventDTO:
      type: object
      properties:
        time:
          type: string
          example: "10:35"
        type:
          type: string
          enum: [SYSTEM, ASSET, ERROR]
        source:
          type: string
          example: "Sensore"
        message:
          type: string